+++ src/Synapse/Synapse/Common/HelpMessages.cs
@@ -426,5 +426,9 @@ For more information, see https://docs.microsoft.com/en-us/sql/relational-databa
         public const string KeyResourceId = "The resource identifier of Synapse SQL Pool.";
 
         public const string KeyObject = "Workspace key input object, usually passed through the pipeline.";
+
+        public const string SparkJobDefinitionName = "The Spark job definition name.";
+
+        public const string SparkJobDefinitionObject = "The Spark job definition object.";
     }
 }