+++ src/Network/Network/ChangeLog.md
@@ -19,12 +19,12 @@
 --->
 
 ## Upcoming Release
-* Added support for viewing extended location of public IP address in the console
-    - `New-AzPublicIpAddress`
-    - `Get-AzPublicIpAddress`
 * Updated cmdlets to enable specification of edge zone
     - `New-AzPublicIpPrefix`
     - `New-AzLoadBalancer`
+* Added support for viewing extended location of public IP address in the console
+    - `New-AzPublicIpAddress`
+    - `Get-AzPublicIpAddress`
 
 ## Version 4.9.0
 * Updated cmdlets for route server for a more stable way to add IP configuration.