+++ src/Network/Network/ChangeLog.md
@@ -19,6 +19,9 @@
 --->
 
 ## Upcoming Release
+* Added support for viewing extended location of virtual network in the console
+    - `New-AzVirtualNetwork`
+    - `Get-AzVirtualNetwork`
 * Added support for viewing extended location of public IP address in the console
     - `New-AzPublicIpAddress`
     - `Get-AzPublicIpAddress`