+++ src/Network/Network/ChangeLog.md
@@ -22,6 +22,9 @@
 * Updated cmdlets to enable specification of edge zone
     - `New-AzPublicIpPrefix`
     - `New-AzLoadBalancer`
+* Added support for viewing extended location of virtual network in the console
+    - `New-AzVirtualNetwork`
+    - `Get-AzVirtualNetwork`
 * Added support for viewing extended location of public IP address in the console
     - `New-AzPublicIpAddress`
     - `Get-AzPublicIpAddress`