+++ src/Network/Network/ChangeLog.md
@@ -19,14 +19,15 @@
 --->
 
 ## Upcoming Release
+* Updated cmdlets to enable specification of edge zone
+    - `New-AzPublicIpPrefix`
+    - `New-AzLoadBalancer`
 * Added support for viewing extended location of virtual network in the console
     - `New-AzVirtualNetwork`
     - `Get-AzVirtualNetwork`
 * Added support for viewing extended location of public IP address in the console
     - `New-AzPublicIpAddress`
     - `Get-AzPublicIpAddress`
-* Updated cmdlets to enable specification of edge zone
-    - `New-AzPublicIpPrefix`
 
 ## Version 4.9.0
 * Updated cmdlets for route server for a more stable way to add IP configuration.